\name{pgPollen-package}
\alias{pgPollen-package}
\alias{pgPollen}
\docType{package}
\title{
\packageTitle{pgPollen}
}
\description{
\packageDescription{pgPollen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pgPollen}
\packageIndices{pgPollen}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pgPollen}

Maintainer: \packageMaintainer{pgPollen}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
